journald: strip whitespace from syslog message only after we forward it to syslogd
In the past people have asked about possibility to preserve all
whitespace in log messages. Our position was that they should configure
forwarding to a traditional syslog implementation. However journald
strips trailing whitespace (part of message body) even before it
forwards message to a syslog daemon. Let's not do and that forward
message to the syslogd as is. Only after that we strstrip() the message
and use stripped message for our own purposes. Note that possible
whitespace characters are still preserved when they appear at the
beginning of a message body.

$ logger "   waldo   "

Before:
======
$ tail /var/log/messages | grep waldo | hexdump -C
00000000  44 65 63 20 31 39 20 31  33 3a 31 39 3a 34 32 20  |Dec 19 13:19:42 |
00000010  6c 6f 63 61 6c 68 6f 73  74 20 76 61 67 72 61 6e  |localhost vagran|
00000020  74 3a 20 20 20 20 77 61  6c 64 6f 0a              |t:    waldo.|
0000002c

After:
======
$ tail /var/log/messages | grep waldo | hexdump -C
00000000  44 65 63 20 31 39 20 31  33 3a 32 31 3a 31 36 20  |Dec 19 13:21:16 |
00000010  6c 6f 63 61 6c 68 6f 73  74 20 76 61 67 72 61 6e  |localhost vagran|
00000020  74 3a 20 20 20 20 77 61  6c 64 6f 20 20 20 0a     |t:    waldo   .|
0000002f

Ansible playbook that I've used to setup rsyslog forwarding can be found
here,

https://gist.github.com/msekletar/7ebc5329962499dd9dbab5a98c4f64ee