core: fix FAMILY_ADDRESS_SIZE() misuse in socket.c
FAMILY_ADDRESS_SIZE() returns the size of in_addr/in6_addr.  It is being
misused in socket.c under the assumption that it returns the size of
sockaddr_in/sockaddr_in6.

Luckily sockaddr_pretty()'s salen argument is unused for
AF_INET/AF_INET6 sockets so the bug is latent.

I have audited the rest of the codebase and found no other instances of
this bug.