shared: fix double free in unmask
Easily reproducible:
1) systemctl mask foo
2) systemctl unmask foo foo

The problem here is that the *i that is put into todo[] is later freed
in strv_uniq(), which is not directly visible from this patch. Somewhere
further in the code, the string that *i pointed to is freed again. That
happens only when multiple services with the same name/path are specified.

On my system, the memory corruption revealed itself a lot later in the
program, which resulted in backtraces that were pretty much misleading
and useless.