systemctl: 'show' don't exit with a failure status if the requested property does not exist
"systemctl show -pUnknown <service>" used to exit with '0' even if the property
passed by '-p' doesn't exist.

But since commit 3dced37b7c2c9a5c733817 (v231+), it exits with a failure
status.

"systemctl show" is supposed to be scriptable and therefore its behavior is
supposed to be stable.

This patch restores the old behavior on which a couple of scripts already rely
now.

Also when the requested property doesn't exist, it always logs it at the debug
level since this part of the code is only used by the show command.

Fixes: #5118